refactor(tape): establish domain ports
