+++ src/main/session/data/index.ts
@@ -1,5 +1,6 @@
 import type { DatabaseConnectionProvider } from '@/data/databaseConnection'
-import type { DeepChatTapeEntryRow } from '@/session/data/tables/deepchatTapeEntries'
+import type { DeepChatTapeEntryRow } from '@/tape/domain/entry'
+import type { TapeMutationProjection } from '@/tape/ports/storage'
 import type { SessionTapePort } from './contracts'
 import { SessionPendingInputStore } from './pendingInputStore'
 import { SessionPendingInputs } from './pendingInputs'
@@ -7,11 +8,10 @@ import { SessionSettingsStore } from './settings'
 import { normalizeTapeHandoffState, SessionTape } from './tape'
 import { SessionTranscript } from './transcript'
 import { SessionDatabase } from './database'
-import type { DeepChatTapeMutationProjection } from './tables/deepchatTapeEntries'
 
 export function createSessionData(
   connection: DatabaseConnectionProvider,
-  getTapeMutationProjection: (() => DeepChatTapeMutationProjection) | undefined,
+  getTapeMutationProjection: (() => TapeMutationProjection) | undefined,
   events: SessionDataEvents
 ) {
   const database = new SessionDatabase(connection, getTapeMutationProjection)
